+++ includes/js/wpglobus-wpseo-22.js
@@ -13,6 +13,10 @@ var wpglobus_wpseo = function () {
 		if (typeof str === 'undefined') {
 			return;
 		}
+		if ( typeof replacedVars === 'undefined' && jQuery('#title').size() == 0 ) {
+            // WP SEO doesn't work properly when post title is disabled
+			return str;	
+		}	
 		var post_title = '#title',
 			post_excerpt = '#excerpt-' + language,
 			post_content = '#content';