+++ wpglobus.php
@@ -5,7 +5,7 @@
  * Description: A WordPress Globalization / Multilingual Plugin. Posts, pages, menus, widgets and even custom fields - in multiple languages!
  * Text Domain: wpglobus
  * Domain Path: /languages/
- * Version: 1.2.0
+ * Version: 1.2.1
  * Author: WPGlobus
  * Author URI: http://www.wpglobus.com/
  * Network: false
@@ -29,7 +29,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'WPGLOBUS_VERSION', '1.2.0' );
+define( 'WPGLOBUS_VERSION', '1.2.1' );
 define( 'WPGLOBUS_PLUGIN_BASENAME', plugin_basename( __FILE__ ) );
 
 /** @todo Get rid of these */
@@ -54,6 +54,13 @@ WPGlobus::Config();
 
 require_once 'includes/class-wpglobus-filters.php';
 require_once 'includes/wpglobus-controller.php';
+
+/**
+ * Fix multilingual strings in `Customize`
+ * @since 1.2.1
+ */
+require_once 'includes/class-wpglobus-customize.php';
+WPGlobus_Customize::controller();
 
 if ( defined( 'WPSEO_VERSION' ) ) {
 	require_once 'includes/class-wpglobus-wpseo.php';